+++ packages/react/src/reality/type.ts
@@ -1,4 +1,4 @@
-import { Vec3 } from '@webspatial/core-sdk'
+import type { Quaternion, Vec3 } from '@webspatial/core-sdk'
 import { EntityRefShape } from './hooks'
 import { SpatialTapEvent as CoreSpatialTapEvent } from '@webspatial/core-sdk'
 import { SpatialDragStartEvent as CoreSpatialDragStartEvent } from '@webspatial/core-sdk'
@@ -28,6 +28,9 @@ export type SpatialTapEntityEvent<T extends EntityRefShape = EntityRefShape> =
       readonly offsetX: number
       readonly offsetY: number
       readonly offsetZ: number
+      readonly clientX: number
+      readonly clientY: number
+      readonly clientZ: number
     }
 
 // drag
@@ -38,10 +41,18 @@ export type SpatialDragStartEntityEvent<
     readonly offsetX: number
     readonly offsetY: number
     readonly offsetZ: number
+    readonly clientX: number
+    readonly clientY: number
+    readonly clientZ: number
   }
 
 export type SpatialDragEntityEvent<T extends EntityRefShape = EntityRefShape> =
-  CoreSpatialDragEvent & allTarget<T>
+  CoreSpatialDragEvent &
+    allTarget<T> & {
+      readonly translationX: number
+      readonly translationY: number
+      readonly translationZ: number
+    }
 
 export type SpatialDragEndEntityEvent<
   T extends EntityRefShape = EntityRefShape,
@@ -49,14 +60,20 @@ export type SpatialDragEndEntityEvent<
 // rotate
 export type SpatialRotateEntityEvent<
   T extends EntityRefShape = EntityRefShape,
-> = CoreSpatialRotateEvent & allTarget<T>
+> = CoreSpatialRotateEvent &
+  allTarget<T> & {
+    readonly quaternion: Quaternion
+  }
 export type SpatialRotateEndEntityEvent<
   T extends EntityRefShape = EntityRefShape,
 > = CoreSpatialRotateEndEvent & allTarget<T>
 // magnify
 export type SpatialMagnifyEntityEvent<
   T extends EntityRefShape = EntityRefShape,
-> = CoreSpatialMagnifyEvent & allTarget<T>
+> = CoreSpatialMagnifyEvent &
+  allTarget<T> & {
+    readonly magnification: number
+  }
 export type SpatialMagnifyEndEntityEvent<
   T extends EntityRefShape = EntityRefShape,
 > = CoreSpatialMagnifyEndEvent & allTarget<T>