+++ apps/test-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web-content
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [fd29643]
+  - @webspatial/react-sdk@1.2.1
+  - @webspatial/core-sdk@1.2.1
+
 ## 0.0.19
 
 ### Patch Changes